feat: per-challenge tsconfig support (#1262)
* feat(db): add tsconfig to challenges

* feat(monaco): respect challenge's tsconfig

This patch also makes `SplitEditor` completely overwrite the TS config on
mount, as it's global. This ensures that it's properly reset for every
challenge.

The default Monaco config doesn't set much: just `allowNonTsExtensions`
and `target`, and we already overwrite the latter. I've added the former
to preserve the defaults.

Reference:
https://github.com/microsoft/monaco-editor/blob/091550e89857bd926d2b66e100dc3c2094475783/src/language/typescript/monaco.contribution.ts#L753

* feat(db): ingest challenge tsconfigs

* feat(db): override tsconfig for "assert array index"

This patch overrides the tsconfig for challenge 925: Assert Array Index,
which requires `noUncheckedIndexedAccess` to be set.

It is theoretically possible to extract the config from the TS playground
URL that tsch.js.org redirects to for each challenge, but even doing a
HEAD request for every challenge appears to be too slow in practice.
Given that this is the only type challenge that requires special config
options as of today, I don't think it makes sense doing that.

* fix: type guard for tsconfig

---------

Co-authored-by: bautistaaa <chrisbautistaaa@gmail.com>